+++ 1. Very Easy/2Sum.cpp
@@ -4,12 +4,13 @@
 
 #include<iostream>
 using namespace std;
+
 int addition(int a, int b) {
 	return a + b;
 }
 
 int main() {
-    int a,b,sum;
+    int a, b, sum;
     cin >> a;
     cin >> b;
     sum =  addition(a,b);
