+++ 1. Very Easy/areaOfTriangle.cpp
@@ -1,18 +1,16 @@
-//Problem:Write a function that takes the base and height of a triangle and return its area.
-//Problem Link:https://edabit.com/challenge/3CaszbdZYGN4otQD8
-#include <iostream>
+//Problem statement :Area of Triangle
+//Probelm Link : https://edabit.com/challenge/HvYiBXgfPtnDHitym
+//Choose one Cateorgy - Very Easy, Easy, Medium, Hard, Very Hard, Expert
+//question category : Very Easy
+
+#include<bits/stdc++.h>
 using namespace std;
-double areaOfTriangle(int h, int b)
-{
-	double area=( (double)h* (double)b ) / 2;
-	return area;
-}
+
 int main()
 {
-	int height,base;
-	cin >> height >> base;
-
-	double area = areaOfTriangle(height,base);
-	cout << area;
-	return 0;
+float base,height;
+cout<<"Enter Base\n"; cin>>base;
+cout<<"Enter Height\n"; cin>>height;
+cout<<"Area of Triangle is\n "<<0.5*base*height;
+return 0;
 }
