--- areaOfTriangle.cpp
@@ -1,16 +1,0 @@
-//Problem statement :Area of Triangle
-//Probelm Link : https://edabit.com/challenge/HvYiBXgfPtnDHitym
-//Choose one Cateorgy - Very Easy, Easy, Medium, Hard, Very Hard, Expert
-//question category : Very Easy
-
-#include<bits/stdc++.h>
-using namespace std;
-
-int main()
-{
-float base,height;
-cout<<"Enter Base\n"; cin>>base;
-cout<<"Enter Height\n"; cin>>height;
-cout<<"Area of Triangle is\n "<<0.5*base*height;
-return 0;
-}
